chore(stablecoin-accordion): rename use-hook file
